+++ features/home/interfaces/PieChartData.ts
@@ -2,6 +2,7 @@ export interface PieChartData {
     value: number;
     color: string;
     text: string;
+    documents:string;
     focused?: boolean;
     percentage?: string; 
 }